pyproject.toml
[project]
name = "convnext-tts"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "reppy4620", email = "ilikeniku@icloud.com" }
]
dependencies = [
    "torch>=2.2.2",
    "torchaudio>=2.2.2",
    "lightning>=2.2.3",
    "joblib>=1.4.0",
    "numba>=0.59.1",
    "hydra-core>=1.3.2",
    "scipy>=1.13.0",
    "soundfile>=0.12.1",
    "pyworld>=0.3.4",
    "pandas>=2.2.2",
    "loguru>=0.7.2",
    "tensorboard>=2.16.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/convnext_tts"]
